change to api route

+++ app/Http/Controllers/Admin/PostController.php
@@ -112,24 +112,6 @@ class PostController extends Controller
     }
 
     /**
-     * Update the specified resource in storage.
-     *
-     * @param int $id post id to change status
-     *
-     * @return array status code
-     */
-    public function changeStatus($id)
-    {
-        $post = Post::find($id);
-        $post->status = !$post->status;
-        $post->save();
-        $data['status'] = (int) $post->status;
-        $data['msg'] = __('post.admin.form.updated');
-        return $data;
-        // return redirect()->route('admin.posts.index');
-    }
-
-    /**
      * Remove the specified resource from storage.
      *
      * @param int $id post id
@@ -139,8 +121,13 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::findOrFail($id);
-        $post->delete();
-        session(['message' => __('post.admin.form.deleted')]);
-        return redirect()->route('admin.posts.index');
+        if ($post) {
+            $post->delete();
+            session(['message' => __('post.admin.form.deleted')]);
+            return redirect()->route('admin.posts.index');
+        } else {
+            session(['message' => __('post.admin.form.id_not_found')]);
+        }
+        
     }
 }
